requirements.txt
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
beautifulsoup4==4.13.3
blinker==1.9.0
CacheControl==0.14.2
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
coloredlogs==15.0.1
cryptography==44.0.2
docling-core==2.23.3
dotenv==0.9.9
fastapi==0.115.11
filelock==3.18.0
firebase==4.0.1
firebase-admin==6.7.0
flatbuffers==25.2.10
fsspec==2025.3.0
gitdb==4.0.12
GitPython==3.1.44
google==3.0.0
google-api-core==2.24.2
google-api-python-client==2.164.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-cloud==0.34.0
google-cloud-core==2.4.3
google-cloud-firestore==2.20.1
google-cloud-storage==3.1.0
google-cloud-vision==3.10.1
google-crc32c==1.7.0
google-resumable-media==2.7.2
googleapis-common-protos==1.69.2
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.14.0
httplib2==0.22.0
huggingface-hub==0.29.3
humanfriendly==10.0
idna==3.10
imageio==2.37.0
Jinja2==3.1.3
jsonref==1.1.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
latex2mathml==3.77.0
lazy_loader==0.4
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mpmath==1.3.0
msgpack==1.1.0
narwhals==1.31.0
networkx==3.3
nudenet==3.4.2
NudeNetClassifier==2.1.1
numpy==2.2.4
onnxruntime==1.21.0
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
packaging==24.2
pandas==2.2.3
pillow==11.1.0
proto-plus==1.26.1
protobuf==5.29.3
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pydeck==0.9.1
Pygments==2.19.1
PyJWT==2.10.1
pyparsing==3.2.1
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2025.1
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rpds-py==0.23.1
rsa==4.9
safetensors==0.5.3
scikit-image==0.25.2
scipy==1.15.2
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.6
starlette==0.46.1
streamlit==1.43.2
sympy==1.13.1
tabulate==0.9.0
tenacity==9.0.0
tifffile==2025.3.13
tokenizers==0.21.1
toml==0.10.2
torch==2.6.0+cu126
torchaudio==2.6.0+cu126
torchvision==0.21.0+cu126
tornado==6.4.2
tqdm==4.67.1
transformers==4.49.0
typer==0.12.5
typing_extensions==4.12.2
tzdata==2025.1
uc==2.2.1
uritemplate==4.1.1
urllib3==2.3.0
uv==0.6.7
uvicorn==0.34.0
watchdog==6.0.0
